// frontend/app/model/user.ts
/*
  - Copyright 2022 Sven Loesekann
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
    You may obtain a copy of the License at
    http://www.apache.org/licenses/LICENSE-2.0
    Unless required by applicable law or agreed to in writing, software
    distributed under the License is distributed on an "AS IS" BASIS,
    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
    See the License for the specific language governing permissions and
    limitations under the License.
*/
export interface UserRequest {
  Username: string;
  Password: string;
  Latitude?: number;
  Longitude?: number;
  SearchRadius?: number;
  PostCode?: number;
  TargetDiesel?: string;
  TargetE10?: string;
  TargetE5?: string;
}

export interface UserResponse {
  Token?: string;
  Message?: string;
  PostCode?: number;
  Uuid?: string;
  Longitude?: number;
  Latitude?: number;
  SearchRadius?: number;
  TargetDiesel?: number;
  TargetE5?: number;
  TargetE10?: number;
}